+++ tomcat/src/main/java/nextstep/Application.java
@@ -6,11 +6,11 @@ import org.slf4j.LoggerFactory;
 
 public class Application {
 
-    private static final Logger log = LoggerFactory.getLogger(Application.class);
+	private static final Logger log = LoggerFactory.getLogger(Application.class);
 
-    public static void main(String[] args) {
-        log.info("web server start.");
-        final var tomcat = new Tomcat();
-        tomcat.start();
-    }
+	public static void main(String[] args) {
+		log.info("web server start.");
+		final var tomcat = new Tomcat();
+		tomcat.start();
+	}
 }
