+++ tomcat/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -1,12 +1,8 @@
 package nextstep.jwp.controller;
 
-import java.io.IOException;
-
 import org.apache.coyote.controller.AbstractController;
-import org.apache.coyote.controller.Controller;
 import org.apache.coyote.http11.http.ContentType;
 import org.apache.coyote.http11.http.HttpHeader;
-import org.apache.coyote.http11.http.HttpMethod;
 import org.apache.coyote.http11.http.HttpRequest;
 import org.apache.coyote.http11.http.HttpResponse;
 import org.apache.coyote.http11.http.HttpStatus;
