+++ tomcat/src/main/java/nextstep/jwp/db/InMemoryUserRepository.java
@@ -1,27 +1,28 @@
 package nextstep.jwp.db;
 
-import nextstep.jwp.model.User;
-
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
+import nextstep.jwp.model.User;
+
 public class InMemoryUserRepository {
 
-    private static final Map<String, User> database = new ConcurrentHashMap<>();
+	private static final Map<String, User> database = new ConcurrentHashMap<>();
 
-    static {
-        final User user = new User(1L, "gugu", "password", "hkkang@woowahan.com");
-        database.put(user.getAccount(), user);
-    }
+	static {
+		final User user = new User(1L, "gugu", "password", "hkkang@woowahan.com");
+		database.put(user.getAccount(), user);
+	}
 
-    public static void save(User user) {
-        database.put(user.getAccount(), user);
-    }
+	public static void save(User user) {
+		database.put(user.getAccount(), user);
+	}
 
-    public static Optional<User> findByAccount(String account) {
-        return Optional.ofNullable(database.get(account));
-    }
+	public static Optional<User> findByAccount(String account) {
+		return Optional.ofNullable(database.get(account));
+	}
 
-    private InMemoryUserRepository() {}
+	private InMemoryUserRepository() {
+	}
 }
