+++ tomcat/src/main/java/nextstep/jwp/exception/UncheckedServletException.java
@@ -2,7 +2,7 @@ package nextstep.jwp.exception;
 
 public class UncheckedServletException extends RuntimeException {
 
-    public UncheckedServletException(Exception e) {
-        super(e);
-    }
+	public UncheckedServletException(Exception e) {
+		super(e);
+	}
 }
