+++ tomcat/src/main/java/nextstep/jwp/model/User.java
@@ -2,37 +2,37 @@ package nextstep.jwp.model;
 
 public class User {
 
-    private final Long id;
-    private final String account;
-    private final String password;
-    private final String email;
+	private final Long id;
+	private final String account;
+	private final String password;
+	private final String email;
 
-    public User(Long id, String account, String password, String email) {
-        this.id = id;
-        this.account = account;
-        this.password = password;
-        this.email = email;
-    }
+	public User(Long id, String account, String password, String email) {
+		this.id = id;
+		this.account = account;
+		this.password = password;
+		this.email = email;
+	}
 
-    public User(String account, String password, String email) {
-        this(null, account, password, email);
-    }
+	public User(String account, String password, String email) {
+		this(null, account, password, email);
+	}
 
-    public boolean checkPassword(String password) {
-        return this.password.equals(password);
-    }
+	public boolean checkPassword(String password) {
+		return this.password.equals(password);
+	}
 
-    public String getAccount() {
-        return account;
-    }
+	public String getAccount() {
+		return account;
+	}
 
-    @Override
-    public String toString() {
-        return "User{" +
-                "id=" + id +
-                ", account='" + account + '\'' +
-                ", email='" + email + '\'' +
-                ", password='" + password + '\'' +
-                '}';
-    }
+	@Override
+	public String toString() {
+		return "User{" +
+			"id=" + id +
+			", account='" + account + '\'' +
+			", email='" + email + '\'' +
+			", password='" + password + '\'' +
+			'}';
+	}
 }
