+++ tomcat/src/main/java/org/apache/catalina/Manager.java
@@ -1,8 +1,8 @@
 package org.apache.catalina;
 
-import jakarta.servlet.http.HttpSession;
-
 import java.io.IOException;
+
+import jakarta.servlet.http.HttpSession;
 
 /**
  * A <b>Manager</b> manages the pool of Sessions that are associated with a
@@ -24,33 +24,33 @@ import java.io.IOException;
  */
 public interface Manager {
 
-    /**
-     * Add this Session to the set of active Sessions for this Manager.
-     *
-     * @param session Session to be added
-     */
-    void add(HttpSession session);
+	/**
+	 * Add this Session to the set of active Sessions for this Manager.
+	 *
+	 * @param session Session to be added
+	 */
+	void add(HttpSession session);
 
-    /**
-     * Return the active Session, associated with this Manager, with the
-     * specified session id (if any); otherwise return <code>null</code>.
-     *
-     * @param id The session id for the session to be returned
-     *
-     * @exception IllegalStateException if a new session cannot be
-     *  instantiated for any reason
-     * @exception IOException if an input/output error occurs while
-     *  processing this request
-     *
-     * @return the request session or {@code null} if a session with the
-     *         requested ID could not be found
-     */
-    HttpSession findSession(String id) throws IOException;
+	/**
+	 * Return the active Session, associated with this Manager, with the
+	 * specified session id (if any); otherwise return <code>null</code>.
+	 *
+	 * @param id The session id for the session to be returned
+	 *
+	 * @exception IllegalStateException if a new session cannot be
+	 *  instantiated for any reason
+	 * @exception IOException if an input/output error occurs while
+	 *  processing this request
+	 *
+	 * @return the request session or {@code null} if a session with the
+	 *         requested ID could not be found
+	 */
+	HttpSession findSession(String id) throws IOException;
 
-    /**
-     * Remove this Session from the active Sessions for this Manager.
-     *
-     * @param session Session to be removed
-     */
-    void remove(HttpSession session);
+	/**
+	 * Remove this Session from the active Sessions for this Manager.
+	 *
+	 * @param session Session to be removed
+	 */
+	void remove(HttpSession session);
 }
