+++ tomcat/src/main/java/org/apache/catalina/connector/Connector.java
@@ -1,94 +1,94 @@
 package org.apache.catalina.connector;
 
-import org.apache.coyote.http11.Http11Processor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.net.ServerSocket;
 import java.net.Socket;
 
+import org.apache.coyote.http11.Http11Processor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 public class Connector implements Runnable {
 
-    private static final Logger log = LoggerFactory.getLogger(Connector.class);
+	private static final Logger log = LoggerFactory.getLogger(Connector.class);
 
-    private static final int DEFAULT_PORT = 8080;
-    private static final int DEFAULT_ACCEPT_COUNT = 100;
+	private static final int DEFAULT_PORT = 8080;
+	private static final int DEFAULT_ACCEPT_COUNT = 100;
 
-    private final ServerSocket serverSocket;
-    private boolean stopped;
+	private final ServerSocket serverSocket;
+	private boolean stopped;
 
-    public Connector() {
-        this(DEFAULT_PORT, DEFAULT_ACCEPT_COUNT);
-    }
+	public Connector() {
+		this(DEFAULT_PORT, DEFAULT_ACCEPT_COUNT);
+	}
 
-    public Connector(final int port, final int acceptCount) {
-        this.serverSocket = createServerSocket(port, acceptCount);
-        this.stopped = false;
-    }
+	public Connector(final int port, final int acceptCount) {
+		this.serverSocket = createServerSocket(port, acceptCount);
+		this.stopped = false;
+	}
 
-    private ServerSocket createServerSocket(final int port, final int acceptCount) {
-        try {
-            final int checkedPort = checkPort(port);
-            final int checkedAcceptCount = checkAcceptCount(acceptCount);
-            return new ServerSocket(checkedPort, checkedAcceptCount);
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
+	private ServerSocket createServerSocket(final int port, final int acceptCount) {
+		try {
+			final int checkedPort = checkPort(port);
+			final int checkedAcceptCount = checkAcceptCount(acceptCount);
+			return new ServerSocket(checkedPort, checkedAcceptCount);
+		} catch (IOException e) {
+			throw new UncheckedIOException(e);
+		}
+	}
 
-    public void start() {
-        var thread = new Thread(this);
-        thread.setDaemon(true);
-        thread.start();
-        stopped = false;
-    }
+	public void start() {
+		var thread = new Thread(this);
+		thread.setDaemon(true);
+		thread.start();
+		stopped = false;
+	}
 
-    @Override
-    public void run() {
-        while (!stopped) {
-            connect();
-        }
-    }
+	@Override
+	public void run() {
+		while (!stopped) {
+			connect();
+		}
+	}
 
-    private void connect() {
-        try {
-            process(serverSocket.accept());
-        } catch (IOException e) {
-            log.error(e.getMessage(), e);
-        }
-    }
+	private void connect() {
+		try {
+			process(serverSocket.accept());
+		} catch (IOException e) {
+			log.error(e.getMessage(), e);
+		}
+	}
 
-    private void process(final Socket connection) {
-        if (connection == null) {
-            return;
-        }
-        log.info("connect host: {}, port: {}", connection.getInetAddress(), connection.getPort());
-        var processor = new Http11Processor(connection);
-        new Thread(processor).start();
-    }
+	private void process(final Socket connection) {
+		if (connection == null) {
+			return;
+		}
+		log.info("connect host: {}, port: {}", connection.getInetAddress(), connection.getPort());
+		var processor = new Http11Processor(connection);
+		new Thread(processor).start();
+	}
 
-    public void stop() {
-        stopped = true;
-        try {
-            serverSocket.close();
-        } catch (IOException e) {
-            log.error(e.getMessage(), e);
-        }
-    }
+	public void stop() {
+		stopped = true;
+		try {
+			serverSocket.close();
+		} catch (IOException e) {
+			log.error(e.getMessage(), e);
+		}
+	}
 
-    private int checkPort(final int port) {
-        final var MIN_PORT = 1;
-        final var MAX_PORT = 65535;
+	private int checkPort(final int port) {
+		final var MIN_PORT = 1;
+		final var MAX_PORT = 65535;
 
-        if (port < MIN_PORT || MAX_PORT < port) {
-            return DEFAULT_PORT;
-        }
-        return port;
-    }
+		if (port < MIN_PORT || MAX_PORT < port) {
+			return DEFAULT_PORT;
+		}
+		return port;
+	}
 
-    private int checkAcceptCount(final int acceptCount) {
-        return Math.max(acceptCount, DEFAULT_ACCEPT_COUNT);
-    }
+	private int checkAcceptCount(final int acceptCount) {
+		return Math.max(acceptCount, DEFAULT_ACCEPT_COUNT);
+	}
 }
