+++ tomcat/src/main/java/org/apache/catalina/startup/Tomcat.java
@@ -1,27 +1,27 @@
 package org.apache.catalina.startup;
 
+import java.io.IOException;
+
 import org.apache.catalina.connector.Connector;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
-
 public class Tomcat {
 
-    private static final Logger log = LoggerFactory.getLogger(Tomcat.class);
+	private static final Logger log = LoggerFactory.getLogger(Tomcat.class);
 
-    public void start() {
-        var connector = new Connector();
-        connector.start();
+	public void start() {
+		var connector = new Connector();
+		connector.start();
 
-        try {
-            // make the application wait until we press any key.
-            System.in.read();
-        } catch (IOException e) {
-            log.error(e.getMessage(), e);
-        } finally {
-            log.info("web server stop.");
-            connector.stop();
-        }
-    }
+		try {
+			// make the application wait until we press any key.
+			System.in.read();
+		} catch (IOException e) {
+			log.error(e.getMessage(), e);
+		} finally {
+			log.info("web server stop.");
+			connector.stop();
+		}
+	}
 }
