+++ tomcat/src/test/java/org/apache/coyote/controller/StaticResourceControllerTest.java
@@ -1,0 +1,75 @@
+package org.apache.coyote.controller;
+
+import static org.assertj.core.api.Assertions.*;
+
+import java.io.File;
+import java.io.IOException;
+import java.net.URL;
+import java.nio.file.Files;
+
+import org.apache.coyote.http11.Http11Processor;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import support.StubSocket;
+
+class StaticResourceControllerTest {
+
+	@DisplayName("/index.html 요청이 들어오면 index.html을 출력한다.")
+	@Test
+	void index() throws IOException {
+		// given
+		final String httpRequest = String.join("\r\n",
+			"GET /index.html HTTP/1.1 ",
+			"Host: localhost:8080 ",
+			"Connection: keep-alive ",
+			"",
+			"");
+
+		final var socket = new StubSocket(httpRequest);
+		final Http11Processor processor = new Http11Processor(socket);
+
+		// when
+		processor.process(socket);
+
+		// then
+		final URL resource = getClass().getClassLoader().getResource("static/index.html");
+		String responseBody = new String(Files.readAllBytes(new File(resource.getFile()).toPath()));
+		var expected = "HTTP/1.1 200 OK \r\n" +
+			"Content-Length: " + responseBody.getBytes().length + " \r\n" +
+			"Content-Type: text/html;charset=utf-8 \r\n" +
+			"\r\n" +
+			responseBody;
+
+		assertThat(socket.output()).isEqualTo(expected);
+	}
+
+	@DisplayName("없는 페이지 요청이 들어오면 404.html을 출력한다.")
+	@Test
+	void notFound() throws IOException {
+		// given
+		final String httpRequest = String.join("\r\n",
+			"GET /notFound.html HTTP/1.1 ",
+			"Host: localhost:8080 ",
+			"Connection: keep-alive ",
+			"",
+			"");
+
+		final var socket = new StubSocket(httpRequest);
+		final Http11Processor processor = new Http11Processor(socket);
+
+		// when
+		processor.process(socket);
+
+		// then
+		final URL resource = getClass().getClassLoader().getResource("static/404.html");
+		String responseBody = new String(Files.readAllBytes(new File(resource.getFile()).toPath()));
+		var expected = "HTTP/1.1 404 Not Found \r\n" +
+			"Content-Length: " + responseBody.getBytes().length + " \r\n" +
+			"Content-Type: text/html;charset=utf-8 \r\n" +
+			"\r\n" +
+			responseBody;
+
+		assertThat(socket.output()).isEqualTo(expected);
+	}
+}
