+++ tomcat/src/test/java/study/FileTest.java
@@ -18,39 +18,37 @@ import org.junit.jupiter.api.Test;
 @DisplayName("File 클래스 학습 테스트")
 class FileTest {
 
-    /**
-     * File 객체를 생성하려면 파일의 경로를 알아야 한다.
-     * 자바 애플리케이션은 resource 디렉터리에 정적 파일을 저장한다.
-     * resource 디렉터리의 경로는 어떻게 알아낼 수 있을까?
-     */
-    @DisplayName("resource_디렉터리에_있는_파일의_경로를_찾는다")
-    @Test
-    void findFilePath() {
-        final String fileName = "nextstep.txt";
+	/**
+	 * File 객체를 생성하려면 파일의 경로를 알아야 한다.
+	 * 자바 애플리케이션은 resource 디렉터리에 정적 파일을 저장한다.
+	 * resource 디렉터리의 경로는 어떻게 알아낼 수 있을까?
+	 */
+	@DisplayName("resource_디렉터리에_있는_파일의_경로를_찾는다")
+	@Test
+	void findFilePath() {
+		final String fileName = "nextstep.txt";
 
-        File file = new File("./src/test/resources/nextstep.txt");
-        final String actual = file.getAbsolutePath();
+		File file = new File("./src/test/resources/nextstep.txt");
+		final String actual = file.getAbsolutePath();
 
-        assertThat(file.exists()).isTrue();
-        assertThat(actual).endsWith(fileName);
-    }
+		assertThat(file.exists()).isTrue();
+		assertThat(actual).endsWith(fileName);
+	}
 
-    /**
-     * 읽어온 파일의 내용을 I/O Stream을 사용해서 사용자에게 전달 해야 한다.
-     * File, Files 클래스를 사용하여 파일의 내용을 읽어보자.
-     */
-    @DisplayName("파일의_내용을_읽는다")
-    @Test
-    void readFile() throws IOException {
-        final String fileName = "nextstep.txt";
+	/**
+	 * 읽어온 파일의 내용을 I/O Stream을 사용해서 사용자에게 전달 해야 한다.
+	 * File, Files 클래스를 사용하여 파일의 내용을 읽어보자.
+	 */
+	@DisplayName("파일의_내용을_읽는다")
+	@Test
+	void readFile() throws IOException {
+		final String fileName = "nextstep.txt";
 
-        // todo
-        File file = new File("./src/test/resources/nextstep.txt");
-        final Path path = file.toPath();
+		File file = new File("./src/test/resources/nextstep.txt");
+		final Path path = file.toPath();
 
-        // todo
-        final List<String> actual = Files.readAllLines(path);
+		final List<String> actual = Files.readAllLines(path);
 
-        assertThat(actual).containsOnly("nextstep");
-    }
+		assertThat(actual).containsOnly("nextstep");
+	}
 }
