+++ tomcat/src/test/java/study/IOStreamTest.java
@@ -1,13 +1,22 @@
 package study;
 
+import static org.assertj.core.api.Assertions.*;
+import static org.mockito.Mockito.*;
+
+import java.io.BufferedInputStream;
+import java.io.BufferedOutputStream;
+import java.io.BufferedReader;
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.FilterInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.io.OutputStream;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-
-import java.io.*;
-
-import static org.assertj.core.api.Assertions.*;
-import static org.mockito.Mockito.*;
 
 /**
  * 입출력(I/O)은 하나의 시스템에서 다른 시스템으로 데이터를 이동 시킬 때 사용한다.
@@ -24,204 +33,203 @@ import static org.mockito.Mockito.*;
 @DisplayName("Java I/O Stream 클래스 학습 테스트")
 class IOStreamTest {
 
-    /**
-     * 자바의 기본 출력 클래스는 java.io.OutputStream이다.
-     * OutputStream의 write(int b) 메서드는 기반 메서드이다.
-     * <code>public abstract void write(int b) throws IOException;</code>
-     */
-    @DisplayName("OutputStream_학습_테스트")
-    @Nested
-    class OutputStreamTest {
+	/**
+	 * 자바의 기본 출력 클래스는 java.io.OutputStream이다.
+	 * OutputStream의 write(int b) 메서드는 기반 메서드이다.
+	 * <code>public abstract void write(int b) throws IOException;</code>
+	 */
+	@DisplayName("OutputStream_학습_테스트")
+	@Nested
+	class OutputStreamTest {
 
-        /**
-         * OutputStream은 다른 매체에 바이트로 데이터를 쓸 때 사용한다.
-         * OutputStream의 서브 클래스(subclass)는 특정 매체에 데이터를 쓰기 위해 write(int b) 메서드를 사용한다.
-         * 예를 들어, FilterOutputStream은 파일로 데이터를 쓸 때,
-         * DataOutputStream은 자바의 primitive type data를 다른 매체로 데이터를 쓸 때 사용한다.
-         * 
-         * write 메서드는 데이터를 바이트로 출력하기 때문에 비효율적이다.
-         * <code>write(byte[] data)</code>와 <code>write(byte b[], int off, int len)</code> 메서드는
-         * 1바이트 이상을 한 번에 전송 할 수 있어 훨씬 효율적이다.
-         */
-        @DisplayName("OutputStream은_데이터를_바이트로_처리한다")
-        @Test
-        void byteProcessing() throws IOException {
-            final byte[] bytes = {110, 101, 120, 116, 115, 116, 101, 112};
-            final OutputStream outputStream = new ByteArrayOutputStream(bytes.length);
+		/**
+		 * OutputStream은 다른 매체에 바이트로 데이터를 쓸 때 사용한다.
+		 * OutputStream의 서브 클래스(subclass)는 특정 매체에 데이터를 쓰기 위해 write(int b) 메서드를 사용한다.
+		 * 예를 들어, FilterOutputStream은 파일로 데이터를 쓸 때,
+		 * DataOutputStream은 자바의 primitive type data를 다른 매체로 데이터를 쓸 때 사용한다.
+		 *
+		 * write 메서드는 데이터를 바이트로 출력하기 때문에 비효율적이다.
+		 * <code>write(byte[] data)</code>와 <code>write(byte b[], int off, int len)</code> 메서드는
+		 * 1바이트 이상을 한 번에 전송 할 수 있어 훨씬 효율적이다.
+		 */
+		@DisplayName("OutputStream은_데이터를_바이트로_처리한다")
+		@Test
+		void byteProcessing() throws IOException {
+			final byte[] bytes = {110, 101, 120, 116, 115, 116, 101, 112};
+			final OutputStream outputStream = new ByteArrayOutputStream(bytes.length);
 
-            /**
-             * OutputStream 객체의 write 메서드를 사용해서 테스트를 통과시킨다
-             */
-            outputStream.write(bytes);
+			/**
+			 * OutputStream 객체의 write 메서드를 사용해서 테스트를 통과시킨다
+			 */
+			outputStream.write(bytes);
 
-            final String actual = outputStream.toString();
+			final String actual = outputStream.toString();
 
-            assertThat(actual).isEqualTo("nextstep");
-            outputStream.close();
-        }
+			assertThat(actual).isEqualTo("nextstep");
+			outputStream.close();
+		}
 
-        /**
-         * 효율적인 전송을 위해 스트림에서 버퍼링을 사용 할 수 있다.
-         * BufferedOutputStream 필터를 연결하면 버퍼링이 가능하다.
-         * 
-         * 버퍼링을 사용하면 OutputStream을 사용할 때 flush를 사용하자.
-         * flush() 메서드는 버퍼가 아직 가득 차지 않은 상황에서 강제로 버퍼의 내용을 전송한다.
-         * Stream은 동기(synchronous)로 동작하기 때문에 버퍼가 찰 때까지 기다리면 데드락(deadlock) 상태가 되기 때문에 flush로 해제한다.
-         */
-        @DisplayName("BufferedOutputStream을_사용하면_버퍼링이_가능하다")
-        @Test
-        void buffering() throws IOException {
-            final OutputStream outputStream = mock(BufferedOutputStream.class);
+		/**
+		 * 효율적인 전송을 위해 스트림에서 버퍼링을 사용 할 수 있다.
+		 * BufferedOutputStream 필터를 연결하면 버퍼링이 가능하다.
+		 *
+		 * 버퍼링을 사용하면 OutputStream을 사용할 때 flush를 사용하자.
+		 * flush() 메서드는 버퍼가 아직 가득 차지 않은 상황에서 강제로 버퍼의 내용을 전송한다.
+		 * Stream은 동기(synchronous)로 동작하기 때문에 버퍼가 찰 때까지 기다리면 데드락(deadlock) 상태가 되기 때문에 flush로 해제한다.
+		 */
+		@DisplayName("BufferedOutputStream을_사용하면_버퍼링이_가능하다")
+		@Test
+		void buffering() throws IOException {
+			final OutputStream outputStream = mock(BufferedOutputStream.class);
 
-            /**
-             * flush를 사용해서 테스트를 통과시킨다.
-             * ByteArrayOutputStream과 어떤 차이가 있을까?
-             */
-            outputStream.flush();
+			/**
+			 * flush를 사용해서 테스트를 통과시킨다.
+			 * ByteArrayOutputStream과 어떤 차이가 있을까?
+			 */
+			outputStream.flush();
 
-            verify(outputStream, atLeastOnce()).flush();
-            outputStream.close();
-        }
+			verify(outputStream, atLeastOnce()).flush();
+			outputStream.close();
+		}
 
-        /**
-         * 스트림 사용이 끝나면 항상 close() 메서드를 호출하여 스트림을 닫는다.
-         * 장시간 스트림을 닫지 않으면 파일, 포트 등 다양한 리소스에서 누수(leak)가 발생한다.
-         */
-        @DisplayName("OutputStream은_사용하고_나서_close_처리를_해준다")
-        @Test
-        void close() throws IOException {
-            final OutputStream outputStream = mock(OutputStream.class);
+		/**
+		 * 스트림 사용이 끝나면 항상 close() 메서드를 호출하여 스트림을 닫는다.
+		 * 장시간 스트림을 닫지 않으면 파일, 포트 등 다양한 리소스에서 누수(leak)가 발생한다.
+		 */
+		@DisplayName("OutputStream은_사용하고_나서_close_처리를_해준다")
+		@Test
+		void close() throws IOException {
+			final OutputStream outputStream = mock(OutputStream.class);
 
-            /**
-             * try-with-resources를 사용한다.
-             * java 9 이상에서는 변수를 try-with-resources로 처리할 수 있다.
-             */
-            try (outputStream) {
-            }
+			/**
+			 * try-with-resources를 사용한다.
+			 * java 9 이상에서는 변수를 try-with-resources로 처리할 수 있다.
+			 */
+			try (outputStream) {
+			}
 
-            verify(outputStream, atLeastOnce()).close();
+			verify(outputStream, atLeastOnce()).close();
 
-        }
-    }
+		}
+	}
 
-    /**
-     * 자바의 기본 입력 클래스는 java.io.InputStream이다.
-     * InputStream은 다른 매체로부터 바이트로 데이터를 읽을 때 사용한다.
-     * InputStream의 read() 메서드는 기반 메서드이다.
-     * <code>public abstract int read() throws IOException;</code>
-     * 
-     * InputStream의 서브 클래스(subclass)는 특정 매체에 데이터를 읽기 위해 read() 메서드를 사용한다.
-     */
-    @DisplayName("InputStream_학습_테스트")
-    @Nested
-    class InputStreamTest {
+	/**
+	 * 자바의 기본 입력 클래스는 java.io.InputStream이다.
+	 * InputStream은 다른 매체로부터 바이트로 데이터를 읽을 때 사용한다.
+	 * InputStream의 read() 메서드는 기반 메서드이다.
+	 * <code>public abstract int read() throws IOException;</code>
+	 *
+	 * InputStream의 서브 클래스(subclass)는 특정 매체에 데이터를 읽기 위해 read() 메서드를 사용한다.
+	 */
+	@DisplayName("InputStream_학습_테스트")
+	@Nested
+	class InputStreamTest {
 
-        /**
-         * read() 메서드는 매체로부터 단일 바이트를 읽는데, 0부터 255 사이의 값을 int 타입으로 반환한다.
-         * int 값을 byte 타입으로 변환하면 -128부터 127 사이의 값으로 변환된다.
-         * 그리고 Stream 끝에 도달하면 -1을 반환한다.
-         */
-        @DisplayName("InputStream은_데이터를_바이트로_읽는다")
-        @Test
-        void readByte() throws IOException {
-            byte[] bytes = {-16, -97, -92, -87};
-            final InputStream inputStream = new ByteArrayInputStream(bytes);
+		/**
+		 * read() 메서드는 매체로부터 단일 바이트를 읽는데, 0부터 255 사이의 값을 int 타입으로 반환한다.
+		 * int 값을 byte 타입으로 변환하면 -128부터 127 사이의 값으로 변환된다.
+		 * 그리고 Stream 끝에 도달하면 -1을 반환한다.
+		 */
+		@DisplayName("InputStream은_데이터를_바이트로_읽는다")
+		@Test
+		void readByte() throws IOException {
+			byte[] bytes = {-16, -97, -92, -87};
+			final InputStream inputStream = new ByteArrayInputStream(bytes);
 
-            /**
-             * inputStream에서 바이트로 반환한 값을 문자열로 어떻게 바꿀까?
-             */
-            OutputStream outputStream = new ByteArrayOutputStream(bytes.length);
+			/**
+			 * inputStream에서 바이트로 반환한 값을 문자열로 어떻게 바꿀까?
+			 */
+			OutputStream outputStream = new ByteArrayOutputStream(bytes.length);
 
-            outputStream.write(inputStream.readAllBytes());
-            final String actual = outputStream.toString();
+			outputStream.write(inputStream.readAllBytes());
+			final String actual = outputStream.toString();
 
-            assertThat(actual).isEqualTo("🤩");
-            assertThat(inputStream.read()).isEqualTo(-1);
-            inputStream.close();
-        }
+			assertThat(actual).isEqualTo("🤩");
+			assertThat(inputStream.read()).isEqualTo(-1);
+			inputStream.close();
+		}
 
-        /**
-         * 스트림 사용이 끝나면 항상 close() 메서드를 호출하여 스트림을 닫는다.
-         * 장시간 스트림을 닫지 않으면 파일, 포트 등 다양한 리소스에서 누수(leak)가 발생한다.
-         */
-        @DisplayName("InputStream은_사용하고_나서_close_처리를_해준다")
-        @Test
-        void close() throws IOException {
-            final InputStream inputStream = mock(InputStream.class);
+		/**
+		 * 스트림 사용이 끝나면 항상 close() 메서드를 호출하여 스트림을 닫는다.
+		 * 장시간 스트림을 닫지 않으면 파일, 포트 등 다양한 리소스에서 누수(leak)가 발생한다.
+		 */
+		@DisplayName("InputStream은_사용하고_나서_close_처리를_해준다")
+		@Test
+		void close() throws IOException {
+			final InputStream inputStream = mock(InputStream.class);
 
-            /**
-             * todo
-             * try-with-resources를 사용한다.
-             * java 9 이상에서는 변수를 try-with-resources로 처리할 수 있다.
-             */
-            try (inputStream) {
-            }
+			/**
+			 * try-with-resources를 사용한다.
+			 * java 9 이상에서는 변수를 try-with-resources로 처리할 수 있다.
+			 */
+			try (inputStream) {
+			}
 
-            verify(inputStream, atLeastOnce()).close();
-        }
-    }
+			verify(inputStream, atLeastOnce()).close();
+		}
+	}
 
-    /**
-     * 필터는 필터 스트림, reader, writer로 나뉜다.
-     * 필터는 바이트를 다른 데이터 형식으로 변환 할 때 사용한다.
-     * reader, writer는 UTF-8, ISO 8859-1 같은 형식으로 인코딩된 텍스트를 처리하는 데 사용된다.
-     */
-    @DisplayName("FilterStream_학습_테스트")
-    @Nested
-    class FilterStreamTest {
+	/**
+	 * 필터는 필터 스트림, reader, writer로 나뉜다.
+	 * 필터는 바이트를 다른 데이터 형식으로 변환 할 때 사용한다.
+	 * reader, writer는 UTF-8, ISO 8859-1 같은 형식으로 인코딩된 텍스트를 처리하는 데 사용된다.
+	 */
+	@DisplayName("FilterStream_학습_테스트")
+	@Nested
+	class FilterStreamTest {
 
-        /**
-         * BufferedInputStream은 데이터 처리 속도를 높이기 위해 데이터를 버퍼에 저장한다.
-         * InputStream 객체를 생성하고 필터 생성자에 전달하면 필터에 연결된다.
-         * 버퍼 크기를 지정하지 않으면 버퍼의 기본 사이즈는 얼마일까?
-         */
-        @DisplayName("필터인_BufferedInWutStream를_사용해보자")
-        @Test
-        void useBufferedIn_OutStream() throws IOException {
-            final String text = "필터에 연결해보자.";
-            final InputStream inputStream = new ByteArrayInputStream(text.getBytes());
-            final InputStream bufferedInputStream = new BufferedInputStream(inputStream);
+		/**
+		 * BufferedInputStream은 데이터 처리 속도를 높이기 위해 데이터를 버퍼에 저장한다.
+		 * InputStream 객체를 생성하고 필터 생성자에 전달하면 필터에 연결된다.
+		 * 버퍼 크기를 지정하지 않으면 버퍼의 기본 사이즈는 얼마일까?
+		 */
+		@DisplayName("필터인_BufferedInWutStream를_사용해보자")
+		@Test
+		void useBufferedIn_OutStream() throws IOException {
+			final String text = "필터에 연결해보자.";
+			final InputStream inputStream = new ByteArrayInputStream(text.getBytes());
+			final InputStream bufferedInputStream = new BufferedInputStream(inputStream);
 
-            final byte[] actual = bufferedInputStream.readAllBytes();
+			final byte[] actual = bufferedInputStream.readAllBytes();
 
-            assertThat(bufferedInputStream).isInstanceOf(FilterInputStream.class);
-            assertThat(actual).isEqualTo("필터에 연결해보자.".getBytes());
-        }
-    }
+			assertThat(bufferedInputStream).isInstanceOf(FilterInputStream.class);
+			assertThat(actual).isEqualTo("필터에 연결해보자.".getBytes());
+		}
+	}
 
-    /**
-     * 자바의 기본 문자열은 UTF-16 유니코드 인코딩을 사용한다.
-     * 바이트를 문자(char)로 처리하려면 인코딩을 신경 써야 한다.
-     * InputStreamReader를 사용하면 지정된 인코딩에 따라 유니코드 문자로 변환할 수 있다.
-     * reader, writer를 사용하면 입출력 스트림을 바이트가 아닌 문자 단위로 데이터를 처리하게 된다.
-     */
-    @DisplayName("InputStreamReader_학습_테스트")
-    @Nested
-    class InputStreamReaderTest {
+	/**
+	 * 자바의 기본 문자열은 UTF-16 유니코드 인코딩을 사용한다.
+	 * 바이트를 문자(char)로 처리하려면 인코딩을 신경 써야 한다.
+	 * InputStreamReader를 사용하면 지정된 인코딩에 따라 유니코드 문자로 변환할 수 있다.
+	 * reader, writer를 사용하면 입출력 스트림을 바이트가 아닌 문자 단위로 데이터를 처리하게 된다.
+	 */
+	@DisplayName("InputStreamReader_학습_테스트")
+	@Nested
+	class InputStreamReaderTest {
 
-        /**
-         * InputStreamReader를 사용해서 바이트를 문자(char)로 읽어온다.
-         * 필터인 BufferedReader를 사용하면 readLine 메서드를 사용해서 문자열(String)을 한 줄 씩 읽어올 수 있다.
-         */
-        @DisplayName("BufferedReader를_사용하여_문자열을_읽어온다")
-        @Test
-        void read() throws IOException {
-            final String emoji = String.join("\r\n",
-                    "😀😃😄😁😆😅😂🤣🥲☺️😊",
-                    "😇🙂🙃😉😌😍🥰😘😗😙😚",
-                    "😋😛😝😜🤪🤨🧐🤓😎🥸🤩",
-                    "");
-            final InputStream inputStream = new ByteArrayInputStream(emoji.getBytes());
+		/**
+		 * InputStreamReader를 사용해서 바이트를 문자(char)로 읽어온다.
+		 * 필터인 BufferedReader를 사용하면 readLine 메서드를 사용해서 문자열(String)을 한 줄 씩 읽어올 수 있다.
+		 */
+		@DisplayName("BufferedReader를_사용하여_문자열을_읽어온다")
+		@Test
+		void read() throws IOException {
+			final String emoji = String.join("\r\n",
+				"😀😃😄😁😆😅😂🤣🥲☺️😊",
+				"😇🙂🙃😉😌😍🥰😘😗😙😚",
+				"😋😛😝😜🤪🤨🧐🤓😎🥸🤩",
+				"");
+			final InputStream inputStream = new ByteArrayInputStream(emoji.getBytes());
 
-            final InputStreamReader inputStreamReader = new InputStreamReader(inputStream);
-            final BufferedReader bufferedReader = new BufferedReader(inputStreamReader);
+			final InputStreamReader inputStreamReader = new InputStreamReader(inputStream);
+			final BufferedReader bufferedReader = new BufferedReader(inputStreamReader);
 
-            final StringBuilder actual = new StringBuilder();
-            actual.append(bufferedReader.readLine()).append("\r\n");
-            actual.append(bufferedReader.readLine()).append("\r\n");
-            actual.append(bufferedReader.readLine()).append("\r\n");
+			final StringBuilder actual = new StringBuilder();
+			actual.append(bufferedReader.readLine()).append("\r\n");
+			actual.append(bufferedReader.readLine()).append("\r\n");
+			actual.append(bufferedReader.readLine()).append("\r\n");
 
-            assertThat(actual).hasToString(emoji);
-        }
-    }
+			assertThat(actual).hasToString(emoji);
+		}
+	}
 }
